Added a new Main class as the program driver
Team decided to redo the driver program in a new class called Main.
This class will execute the input loop needed to run the program.
It asks for user command, checks for validity, executes the command
and then asks for the next command. Also has a welcome and exit message.