RENATT and DELREL bugs fixed and formatting update
There was a bug that allowed a renamed Attribute to be named the
same as an already existing Attribute. Added a check on the new
name entered. Eliminated a bug where after a class was deleted, it
would still appear in relationships of other classes. Now when classes
are deleted a search for relationships that it is the destination of
is done and is removed if found. Other formatting changes made to
output strings.