Field methods and commands updated to MVC design
I changed the methods in Class.java dealing with fields to be compliant
with the MVC design. I added a funtion in Class.java to return a Field
when given a String to simplify the process of determining if a Field
exists within a Class. I also added to View.java methods that are more
general with the print statements so they can be called in many
different scenarios. Like all Renaming and Retyping situations.