Finished(?) GUI
Hey Guys. I have here a 99% finished GUI!. However, I think we have a problem. I can't access any of the controller functions from the popup windows. I believe it is because of how I am setting a passed-in controller to the controller in a window.  Dont worry! I think I have an easy fix. What I am going to do is take the code in the window classes that make the window and put it under its corresponding button in GUIview. I tested that out and it works just fine.