Class buttons working
I have got the buttons to add, delete, and rename classes working. I ended up not using the controller in GUIView instead just calling functions in the model. (kind of redundant tp call the controller to perform model functions when I can just do that from the GUI. Also I added a function in model to return a list of the names of all of the classes in the model. (in order to put those names in the combo boxes) There is a function towards the bottom of the GUIView function called "getList" that right now returns a list of classes in the model, but I am working on making options to get Field and MEthods, among other things. THere are also functions to make small pop -up windows to notify of actions (Class added, class deleted, etc.). Pretty easy to make another one, just copy-paste and change text to whatever you need it to be.