Reworked listcl to have all Class elements
Class contains new elements that need to be printed when ListCLass is
called. Also Attributes needed to be removed and replaced by Fields and
Methods. I added loops to get these new elements and print them neatly.
Parameters had to be formatted to include a comma if there was more than
one Parameter and not include a comma on the last element.