Major bug fixes for Relationship in CLI and GUI
There was problems with how RELTYPE was implemented which caused
all Relationships to be edited to the new type instead of just one.
These changes cause a domino effect in bug fixes in all the files
dealing with Relationships. Implemented checks for blank user input with
error messaging. Fixed some command descriptions in Help text. Converted
Save/Load/Exit to follow MVC. MTDTYPE changed the method's name not
type. In the GUI DELFLD was missing the contents of its combobox.
Added functionality to the cancel button in save. Changed the
DEFAULT.json to JAMBL.json which is actually what is used. Fields were
able to be renamed to existing fields and is now corrected. Parameter
comboboxes need work still. Parameter functionality couldnt be tested.