Addition of Method JUnit Tests and minor bug fixes
Implemented tests for Add, Delete, Rename and Retype for Methods. There
were problems with the existing tests for Relationships that needed to
be fixed due to case sensitivity and spelling errors. The test
deleteClass2 was always always going to fail because the deleteClass
function in the Model was always returning true, so that needed to be
rewritten in Model.