Implementation of Maven build system
The project was converted to comply with Maven to create a build system
that creates a reliable setup for the user to run the program. Java
files needed to be rearranged and the README was updated with the new
requirements and and steps to build and run the program. There was an
error in one of the JUnit tests that was fixed and the Main.java was
updated to take in the --cli argument to change which mode is being used