Getting 'something' to appear
I have made something appear. In the GUIView I have changed the textArea into a JPanel which allowed me to more easily add in components.  In the Class class I have made a method to return a Small JTextArea for that class (right now it just has a name and a dividing line, but will add more) I made it a text area instead of a label because the label did not really let me do any formatting of the string. In GUIController I added a method to get a class's box to be sent to the view