Adding MVC folders and adjestments to the build
I spearated the Java files into MVC folders to keep things more
organized. All the files needed to be included in the package and import
ones as needed. I updated the pom.xml file to adapt to these changes.
Many of the class variables for GUIView needed to be made public to
allow for it to compile. I fixed a bug where the hash set of models
was being accessed directly instead of using the getter.