I got something to drag!
Ok, there's a lot to unpack here. I made a new "draggableBox" class to represent a dragable class box. This draggable box is and extension of a JTextArea I have changed the return type of the getBox function in Class.java to return this new class. There is a getPosition method in the draggableBox class for those who need that sort of information. I have also added the functionality of removing a class and having its corresponding box removed from the display area. Also added the beginning of some miscellaneous methods for changing the contents of the boxes.