Implementation of undo/redo functions
In order to implement the undo/redo capabilities I created a Memento
class that holds a snapshot of the model. To handle these Mementos I
created a History class that performs the undo/redo commands utlilizing
two stacks. I integrated these classes into the Controller and
GUIController by creating a Memento before each command is successfully
executed. Then when an undo or redo command occurs the History can
retrieve the appropriate Memento and load it to the current model.