Draggables updated
A worrisome problem i noticed: I realized the class isn't getting it's box's location anymore. The box is stationary in the MyFrame now while to MyFrame is moving around. There is also a bug where when I change a box's contents it leaves behind the old frame. All the frame stuff is in the jamblPanel.java class