Added a refresh method for the diagramArea.
I modified the MyFrame classboxes to be updated upon every change
made to the Model. During each refresh the new positions of the class
boxes are recorded to the Model and then they are removed from the
diagramArea in order to make sure no duplicates appear. Then all new
MyFrame classboxes are created using the updated Model that now has the
new XY coordinates and whatever change was made from the user command.
Refering to the model everytime the display area is populated ensures
the most recent  data is being shown to the user.

Changes were made to the x and y in Class.java to make them integers to
allow for type matching in frame coordinates. Load.java was also updated
to reflect the int type change.