Integrated the Relationship arrow drawing
I was able to get the lines to draw by moving our previous attempt
into the paintComponent override in jamblPanel.java. I added
a ArrayList of strings into each frame to be interpreted as the
destination and types of each relationship. Sill need to work on
getting the line origin to draw from the center of the box and end at
the middle of the destination side.