Added more tests to increase code coverage
I added tests for parts of the model that werent being covered yet.
I reworked the deleteAllParameters method since it could never return
false and therefore that outcome couldn't be tested for. Some methods
were moved out of the Model files and into the controller to adhere
to the MVC design pattern. I moved the dragglebox.java into the View
as well. This changed some of the imports on other files.