Added classbox autosizing
I implemented classbox autosizing by calculating the longest length of
field or method names that are listed in the class. I also established
an appropriate size for an empty classbox and if the contents does not
exceed that standard size it does not change. I separated the the fields
and methods into their own boxes to better reflect the standard UML
class diagrams. Classboxes also grow in height as more fields and methods
are added to the classbox. I added 2 new parameters to MyFrame in order
to be able to set the size based on the contents.